+++ packages/db/CHANGELOG.md
@@ -1,4 +1,6 @@
 # Changelog
 
+## 0.0.2 - 16/12/2020
+- Added loadFromCache and loadFromArrayBuffer functions
 ## 0.0.1 - 16/12/2020
 - First Release